+++ W40K_addon/W40K_List.py
@@ -1,4 +1,4 @@
-from W40K_UnitCreator import *
+from W40K_Unit import *
 
 KrakGrenade = Weapon(8,2,"Assaut")
 KrakGrenade.set_Quantity(100)
